requirements.txt
addict==2.1.3
amqp==2.3.2
APScheduler==3.0.0
argon2-cffi==18.1.0
bcrypt==3.1.4
bibtexparser==1.0.1
billiard==3.5.0.4
celery==4.2.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.39.1
django-appconf==1.0.3
-e git+https://github.com/yourlabs/django-autocomplete-light.git@13529dd158769c3ea29b6c7bade9d06cb70acf34#egg=django_autocomplete_light
django-bootstrap3==11.0.0
django-bower==5.2.0
django-braces==1.13.0
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extensions==2.1.6
django-extra-views==0.12.0
django-floppyforms==1.7.0
django-grappelli==2.12.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==3.1.2
django-nested-admin==3.2.2
django-notifications-hq==1.5.0
django-ratelimit==2.0.0
django-reversion==3.0.3
django-select2==7.0.3
django-select2-forms==2.0.2
django-sortedm2m==1.5.0
django-tagulous==0.14.0
djangorestframework==3.10.3
dparse==0.4.1
Faker==0.8.12
future==0.16.0
google-scholar-scraper==0.2
gunicorn==19.9.0
habanero==0.6.2
idna==2.6
jsonfield==2.0.2
kombu==4.2.1
misaka==2.1.0
nodeenv==1.3.0
oauthlib==2.1.0
packaging==18.0
Pillow==5.0.0
psycopg2==2.7.7
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.7.0
python-monkey-business==1.0.0
python3-openid==3.1.0
pytz==2018.3
PyYAML==4.2b1
requests==2.22.0
requests-oauthlib==1.0.0
safety==1.8.4
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
tzlocal==2.0.0
urllib3==1.25.6
vine==1.1.4
virtualenv==16.7.6
virtualenv-clone==0.5.3
whitenoise==3.3.0